backend/content/serializers.py
# backend/content/serializers.py

from rest_framework import serializers
from .models import Article

class ArticleSerializer(serializers.ModelSerializer):
    author = serializers.StringRelatedField(read_only=True)

    class Meta:
        model = Article
        fields = ['id', 'title', 'content', 'author', 'published_at']
