backend/requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
distro==1.9.0
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.5.0
kombu==5.4.2
openai==1.51.0
packaging==24.1
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
